File reorg, extract util functions, linting
