+++ dict.json
@@ -1,0 +1,1 @@
+{"filter": {"delete": ["moist", "kaboodle"], "warn": ["Warning", "Questionable Words", "Edgy Potatoes", "turcot"]}}
