+++ kaztron/__init__.py
@@ -1,0 +1,7 @@
+"""
+File description.
+
+Project: Project Name
+"""
+
+__author__ = 'marcalexc'
