+++ kaztron/config.py
@@ -7,6 +7,7 @@ import copy
 
 logger = logging.getLogger("kaztron.config")
 
+
 class KaztronConfig:
     """
     Simple interface for KazTron configuration files. This class uses JSON as
@@ -59,7 +60,7 @@ class KaztronConfig:
         """
         logger.info("config({}) Writing file...".format(self.filename))
         with open(self.filename, "w") as cfg_file:
-            d = json.dump(self._data, cfg_file)
+            json.dump(self._data, cfg_file)
 
     def get_section(self, section: str):
         """
@@ -145,7 +146,7 @@ class KaztronConfig:
 
         try:
             section_data = self._data[section]
-        except KeyError as e:
+        except KeyErrore:
             logger.debug("Section {!r} not found: creating new section".format(section))
             section_data = self._data[section] = {}
 
@@ -157,122 +158,28 @@ def log_level(value: str):
     Converter for KaztronConfig.get() for the core.log_level config
     """
     log_level_map = {
-        'CRITICAL' : logging.CRITICAL,
-        'ERROR' : logging.ERROR,
-        'WARNING' : logging.WARNING,
-        'INFO' : logging.INFO,
-        'DEBUG' : logging.DEBUG,
+        'CRITICAL': logging.CRITICAL,
+        'ERROR': logging.ERROR,
+        'WARNING': logging.WARNING,
+        'INFO': logging.INFO,
+        'DEBUG': logging.DEBUG,
     }
-    return log_level_map.get(value.upper(), DEFAULTS['LogLevel'])
+    return log_level_map[value.upper()]
 
 
-def make_kaztron_config(defaults=None):
-    return KaztronConfig(defaults=defaults)
-
-def make_filter_config():
-    return KaztronConfig("dict.json", defaults={"filter": {"warn": [], "delete": []}})
+_kaztron_config = None
+_filter_config = None
 
 
-DEFAULTS = { 'LogLevel': logging.WARNING, 'LogFile': 'kaztron.log' }
+def get_kaztron_config(defaults=None):
+    global _kaztron_config
+    if not _kaztron_config:
+        _kaztron_config = KaztronConfig(defaults=defaults)
+    return _kaztron_config
 
-token = ""
-modteam = []
-filterdelete = []
-filterwarn = []
-warnchannel = ""
-deletechannel = ""
-outputchannel = ""
-welcomechannel = ""
-ruleschannel = ""
-dicechannel = ""
-testchannel = ""
-authorID = ""
-showcase = ""
-showcase_spreadsheet_id = ""
-showcase_spreadsheet_range = ""
-loglevel = DEFAULTS['LogLevel']
-logfile = DEFAULTS['LogFile']
 
-data = {}
-
-# TODO: Seriously? Refactor this to be a data class with I/O methods...
-def data_import():
-    """
-    :raise OSError: Failure to read config.json
-    """
-    # pls refactor me. end the pain.
-    global token, modteam, filterdelete, filterwarn, warnchannel, \
-           outputchannel, welcomechannel, ruleschannel, dicechannel, \
-           testchannel, authorID, showcase, \
-           showcase_spreadsheet_id, showcase_spreadsheet_range, \
-           loglevel, logfile
-
-    log_level_map = {
-        'CRITICAL' : logging.CRITICAL,
-        'ERROR' : logging.ERROR,
-        'WARNING' : logging.WARNING,
-        'INFO' : logging.INFO,
-        'DEBUG' : logging.DEBUG,
-    }
-
-    # read config file
-    with open("config.json") as json_data:
-        config_data = json.load(json_data)
-
-    # prevalidate and store data
-    token = config_data["token"]
-    modteam = config_data["modteam"]
-    warnchannel = config_data["WarnChannel"]
-    outputchannel = config_data["OutputChannel"]
-    welcomechannel = config_data["WelcomeChannel"]
-    ruleschannel = config_data["RulesChannel"]
-    dicechannel = config_data["DiceChannel"]
-    testchannel = config_data["TestChannel"]
-    authorID = config_data["AuthorID"]
-    showcase = config_data["ShowcaseChannel"]
-    showcase_spreadsheet_id = config_data["ShowcaseSpreadsheetId"]
-    showcase_spreadsheet_range = config_data["ShowcaseSpreadsheetRange"]
-
-    loglevel = log_level_map.get(config_data["LogLevel"].upper(), DEFAULTS['LogLevel'])
-    logfile = config_data["LogFile"] if config_data["LogFile"] else DEFAULTS['LogFile']
-
-    # for API compatibility w/rest of kaztron for now - until a proper refactor
-    return token, modteam, filterdelete, filterwarn, warnchannel, \
-           outputchannel, welcomechannel, dicechannel, testchannel, authorID, \
-           showcase
-
-def data_dump(data, path):
-    with open(path,"w") as json_data:
-        d = json.dump(data,json_data)
-
-def dict_import() -> None:
-    """
-    Import message filter rules (warn/auto-delete) from `dict.json`. If
-    file does not exist, assumes blank rules.
-
-    :raise OSError: Failure to read `dict.json` (except file-not-found, which is
-    handled internally)
-    """
-    global filterdelete, filterwarn
-    try:
-        with open("dict.json") as json_data:
-            dict_data = json.load(json_data)
-    except IOError as e: # file doesn't exist?
-        if e.errno == errno.ENOENT:
-            logger.warn("dict.json does not exist: assuming empty dict")
-            dict_data = {}
-        else:
-            raise
-
-    filterdelete = dict_data.get("delete", [])
-    filterwarn = dict_data.get("warn", [])
-
-def dict_dump(delete, warn):
-    data_dump({"delete" : delete, "warn" : warn}, "dict.json")
-
-def dict_write() -> (list, list):
-    """ Write the message filter rules (warn/auto-delete) to `dict.json`. """
-    global filterdelete, filterwarn
-    dict_dump(filterdelete, filterwarn)
-    dict_import() # reload as written
-    return filterdelete, filterwarn
+def get_filter_config():
+    global _filter_config
+    if not _filter_config:
+        _filter_config = KaztronConfig("dict.json", defaults={"filter": {"warn": [], "delete": []}})
+    return _filter_config
