+++ kaztron/errors.py
@@ -1,0 +1,11 @@
+from discord.ext import commands
+
+
+class UnauthorizedUserError(commands.CommandError):
+    pass
+
+
+class ModOnlyError(UnauthorizedUserError):
+    pass
+
+
