+++ kaztron/wordfilter.py
